requirements.txt
Flask==1.1.1
gunicorn==20.0.4
pytest==5.4.1
python-jenkins==1.7.0
requests==2.23.0
responses==0.10.12
